Moved the autostart setting handling into a separate module.
